[FIX] account_move_cancel_confirm: remove assert related to payment state in test
This assert is not related to the module, because this module doesn’t change anything related to the payment state, and this assert becomes problematic, so we removed it.

Technically: when a vendor bill is paid, the payment state remains in_process instead of paid. This occurs because in _compute_state, when considering whether all invoices are paid, it uses the field reconciled_invoice_ids, which only takes sales documents.

See
https://github.com/odoo/odoo/blob/1cac54db8634267a780b4011291f1e8a80ac5f5b/addons/account/models/account_payment.py#L428
https://github.com/odoo/odoo/blob/1cac54db8634267a780b4011291f1e8a80ac5f5b/addons/account/models/account_payment.py#L699